leetcode-agent-backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
httpx
pydantic
sqlalchemy
psycopg2-binary
cryptography
PyJWT